Fixed some compiler warnings

+++ cocos/3d/CCAnimate3D.cpp
@@ -110,8 +110,8 @@ void Animate3D::startWithTarget(Node *target)
     _boneCurves.clear();
     auto skin = sprite->getSkeleton();
     bool hasCurve = false;
-    for (unsigned int  i = 0; i < skin->getBoneCount(); i++) {
-        auto bone = skin->getBoneByIndex(i);
+    for (int  i = 0; i < skin->getBoneCount(); i++) {
+        auto bone = skin->getBoneByIndex(static_cast<unsigned int>(i));
         auto curve = _animation->getBoneCurveByName(bone->getName());
         if (curve)
         {
